Format broad kernel subsystems under a truncated diff
The T1 diff is truncated, so this is described as a formatting pass across CPU, disk, install, NVMe, ACPI, and related kernel files.

The T1 diff for this commit is truncated, so this message keeps the claim at the file/stat level instead of pretending the internal rewrite is fully visible. This is integration pressure around a networked interactive runtime: AI HTTPS, timers, DNS/adapters, HID, Pixi, boot benchmarks, and faster display paths are made to coexist.

+++ src/cpu.rs
@@ -9,8 +9,8 @@ const AP_HEARTBEAT_TASK_POOL: usize = 256;
 #[unsafe(no_mangle)]
 pub unsafe extern "C" fn ap_start(cpu: &LimineCpu) -> ! {
     enable_sse();
-    let slot = percpu::slot_for_lapic_id(cpu.lapic_id as u32);
-    percpu::init_ap(cpu.lapic_id as u32, slot as u32);
+    let slot = percpu::slot_for_lapic_id(cpu.lapic_id);
+    percpu::init_ap(cpu.lapic_id, slot as u32);
     let ex = percpu::init_executor();
     let spawner = ex.spawner();
